Allows Navigation options to change as userRole is changed
